product model and route part 1

+++ app/services/userService.js
@@ -149,7 +149,11 @@ const uploadImageUser = async (filename, userId) => {
   ) {
     // changing picture
     const oldImagePath = `./public/${user.imageProfilUrl}`;
-    fs.unlinkSync(oldImagePath);
+    if (fs.existsSync(oldImagePath)) {
+      
+      fs.unlinkSync(oldImagePath);
+    }
+    
   }
 
   const newUser = Object.assign(user, {
